fix: polling job runs at the right time
Restore the old hook implementation with some additional conditions.
